+++ institution/views.py
@@ -9,7 +9,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 
 from core import constant
 from core.export_data import cell_values, download_csv_utils
-from core.helper import renderer_utils, query_utils, view_utils, role_utils
+from core.helper import renderer_utils, query_utils, view_utils, perm_utils
 from core.helper.common_recref_adapter import RecrefFormAdapter
 from core.helper.model_utils import ModelLike
 from core.helper.recref_handler import ImageRecrefHandler, TargetResourceFormsetHandler
@@ -149,7 +149,7 @@ def full_form(request, pk):
 
     is_save_success = False
     if request.POST:
-        role_utils.validate_permission_denied(request.user, constant.PM_CHANGE_INST)
+        perm_utils.validate_permission_denied(request.user, constant.PM_CHANGE_INST)
 
         if view_utils.any_invalid_with_log([
             inst_form,
