+++ location/views.py
@@ -13,7 +13,7 @@ from core import constant
 from core.constant import REL_TYPE_COMMENT_REFERS_TO, REL_TYPE_WAS_SENT_TO, REL_TYPE_WAS_SENT_FROM
 from core.export_data import download_csv_utils, cell_values
 from core.forms import CommentForm
-from core.helper import view_utils, renderer_utils, query_utils, role_utils
+from core.helper import view_utils, renderer_utils, query_utils, perm_utils
 from core.helper.common_recref_adapter import RecrefFormAdapter
 from core.helper.model_utils import ModelLike
 from core.helper.recref_handler import RecrefFormsetHandler, ImageRecrefHandler, TargetResourceFormsetHandler
@@ -131,7 +131,7 @@ def full_form(request, location_id):
 
     is_save_success = False
     if request.method == 'POST':
-        role_utils.validate_permission_denied(request.user, constant.PM_CHANGE_LOCATION)
+        perm_utils.validate_permission_denied(request.user, constant.PM_CHANGE_LOCATION)
 
         if fhandler.is_invalid():
             return fhandler.render_form(request)
